requirements.txt
Flask-RESTful
python-dotenv
Flask-JWT-Extended
Flask-SQLAlchemy
pymysql
flask-marshmallow
marshmallow-sqlalchemy
marshmallow
requests
flask-uploads